separate evaluation and compilation phase
For more efficient sampling, a gradient function is created for each parameter.  This obviates the need to lookup values while evaluating samples.

To be able to bring in optimization (e.g. of hyperparameters) more easily, gradients are with respect to Parameters now.  I.e. no longer just to (random) Variables.

Work in progress - models/inference still needs to be converted.